requirements.txt
requests
pydantic
python-dotenv
ortools
fastapi
uvicorn
jinja2
python-multipart
